Consolidate SipHash-2-4 implementation
