+++ DnsServerCore/Dns/Security/DnsCookieValidator.cs
@@ -288,81 +288,5 @@ namespace DnsServerCore.Dns.Security
         }
 
         #endregion
-
-        internal static class SipHash24
-        {
-            // SipHash-2-4 with 128-bit key (16 bytes), returns 64-bit tag.
-            public static ulong Compute(ReadOnlySpan<byte> key16, ReadOnlySpan<byte> msg)
-            {
-                if (key16.Length != 16)
-                    throw new ArgumentException("SipHash key must be 16 bytes.", nameof(key16));
-
-                ulong k0 = ReadU64LE(key16.Slice(0, 8));
-                ulong k1 = ReadU64LE(key16.Slice(8, 8));
-
-                ulong v0 = 0x736f6d6570736575UL ^ k0;
-                ulong v1 = 0x646f72616e646f6dUL ^ k1;
-                ulong v2 = 0x6c7967656e657261UL ^ k0;
-                ulong v3 = 0x7465646279746573UL ^ k1;
-
-                int len = msg.Length;
-                int end = len & ~7;
-
-                for (int i = 0; i < end; i += 8)
-                {
-                    ulong m = ReadU64LE(msg.Slice(i, 8));
-                    v3 ^= m;
-                    SipRound(ref v0, ref v1, ref v2, ref v3);
-                    SipRound(ref v0, ref v1, ref v2, ref v3);
-                    v0 ^= m;
-                }
-
-                ulong b = (ulong)len << 56;
-                int rem = len - end;
-                if (rem != 0)
-                {
-                    ReadOnlySpan<byte> tail = msg.Slice(end, rem);
-                    for (int i = 0; i < rem; i++)
-                        b |= (ulong)tail[i] << (8 * i);
-                }
-
-                v3 ^= b;
-                SipRound(ref v0, ref v1, ref v2, ref v3);
-                SipRound(ref v0, ref v1, ref v2, ref v3);
-                v0 ^= b;
-
-                v2 ^= 0xff;
-                SipRound(ref v0, ref v1, ref v2, ref v3);
-                SipRound(ref v0, ref v1, ref v2, ref v3);
-                SipRound(ref v0, ref v1, ref v2, ref v3);
-                SipRound(ref v0, ref v1, ref v2, ref v3);
-
-                return v0 ^ v1 ^ v2 ^ v3;
-            }
-
-            private static void SipRound(ref ulong v0, ref ulong v1, ref ulong v2, ref ulong v3)
-            {
-                v0 += v1; v1 = RotL(v1, 13); v1 ^= v0; v0 = RotL(v0, 32);
-                v2 += v3; v3 = RotL(v3, 16); v3 ^= v2;
-                v0 += v3; v3 = RotL(v3, 21); v3 ^= v0;
-                v2 += v1; v1 = RotL(v1, 17); v1 ^= v2; v2 = RotL(v2, 32);
-            }
-
-            private static ulong RotL(ulong x, int b) => (x << b) | (x >> (64 - b));
-
-            private static ulong ReadU64LE(ReadOnlySpan<byte> s)
-            {
-                // SipHash spec uses little-endian loads for message words.
-                return
-                    ((ulong)s[0]) |
-                    ((ulong)s[1] << 8) |
-                    ((ulong)s[2] << 16) |
-                    ((ulong)s[3] << 24) |
-                    ((ulong)s[4] << 32) |
-                    ((ulong)s[5] << 40) |
-                    ((ulong)s[6] << 48) |
-                    ((ulong)s[7] << 56);
-            }
-        }
     }
 }
